requirements.txt
fastapi== 0.75.1
uvicorn==0.17.6
tortoise-orm==0.19.0
starlette
simple-crypt==4.1.7
python-jose==3.3.0
passlib==1.7.4
